+++ docs/CHANGES.rst
@@ -5,6 +5,8 @@ Version 0.2.0 (not yet released)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 - added ``remove`` parameter to :py:func:`~iteration_utilities.clamp`.
+- made :py:func:`~iteration_utilities.deepflatten` string-aware. For other
+  recusive-iterable classes a RecursionException is raised instead of freezing.
 
 
 Version 0.1.0 (25.1.2017)